Support building Dex form LLVM HEAD
This is available via the new `llvm-head` Cabal flag which should always
build with the `llvm-head-unstable` branch of `llvm-hs`.
