+++ stack-llvm-head.yaml
@@ -11,7 +11,7 @@ packages:
 
 extra-deps:
   - github: llvm-hs/llvm-hs
-    commit: 442bc488c39f0264930c95e2c98b5cf055d53e8e
+    commit: 0eab4eb0f50949b84d4a66e11be97b335c3532cc
     subdirs:
       - llvm-hs
       - llvm-hs-pure
@@ -20,3 +20,4 @@ extra-deps:
   - store-0.7.8@sha256:0b604101fd5053b6d7d56a4ef4c2addf97f4e08fe8cd06b87ef86f958afef3ae,8001
   - store-core-0.4.4.4@sha256:a19098ca8419ea4f6f387790e942a7a5d0acf62fe1beff7662f098cfb611334c,1430
   - th-utilities-0.2.4.1@sha256:b37d23c8bdabd678aee5a36dd4373049d4179e9a85f34eb437e9cd3f04f435ca,1869
+
